backend/requirements.txt
flask
flask-cors
faster-whisper
openai
python-dotenv
elevenlabs
